Update README.md
Made a few updates for grammar/punctuation/etc.  Also noticed that we were referring to Apache 2.0 license, so changed that to MIT.

I'm also seeing a problem where the rendered file shows step 2 as restarting the numbering so it shows as a "1." again, but I don't know how to fix that in the file here.